+++ src/main/java/nextstep/subway/ControllerExceptionHandler.java
@@ -1,6 +1,7 @@
 package nextstep.subway;
 
-import nextstep.subway.line.exception.NoResourceException;
+import nextstep.subway.common.exception.NoResourceException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -14,5 +15,11 @@ public class ControllerExceptionHandler {
   @ExceptionHandler(NoResourceException.class)
   public ResponseEntity handleNoResourceException(NoResourceException e) {
     return ResponseEntity.notFound().build();
+  }
+
+  @ResponseStatus(HttpStatus.BAD_REQUEST)
+  @ExceptionHandler(DataIntegrityViolationException.class)
+  public ResponseEntity handleIllegalArgsException(DataIntegrityViolationException e) {
+    return ResponseEntity.badRequest().build();
   }
 }
