+++ src/main/java/nextstep/subway/common/exception/NoResourceException.java
@@ -1,8 +1,9 @@
-package nextstep.subway.line.exception;
+package nextstep.subway.common.exception;
 
 public class NoResourceException extends RuntimeException{
 
   public NoResourceException(String message) {
     super(message);
   }
+
 }
